+++ src/Choreography.hs
@@ -1,19 +1,25 @@
-{-# LANGUAGE ExplicitNamespaces #-}
-
 module Choreography (
-  module Choreography.Choreo,
   module Choreography.Core,
-  module Choreography.Location,
+  module Choreography.Choreography,
+  module Choreography.Choreography.Batteries,
+  module Choreography.Locations,
+  module Choreography.Locations.Batteries,
+  module Choreography.Network,
+  module Choreography.Polymorphism,
 
   -- * Running choreographies
   runChoreography
   ) where
 
+import Control.Monad.IO.Class (MonadIO)
+
 import Choreography.Core
-import Choreography.Location
-import Choreography.Choreo
+import Choreography.Choreography
+import Choreography.Choreography.Batteries
+import Choreography.Locations
+import Choreography.Locations.Batteries
 import Choreography.Network
-import Control.Monad.IO.Class
+import Choreography.Polymorphism
 
 -- | Run a choreography with a message transport backend.
 runChoreography :: (Backend config, MonadIO m, KnownSymbols ps) => config -> Choreo ps m a -> LocTm -> m a
